move course header logic to CourseHeaderDelegate
